Reorganize the source code
Adopt the code layout specified in SQR-072. This mostly involves
moving various top-level modules into services or storage directories
as appropriate and updating all of the imports.

Rename mobu.utils to mobu.asyncio on the grounds that this is a set
of utility functions for asyncio and avoiding modules named utils
(the name isn't specific enough to be meaningful) is a good idea.

Rename the templates directory to data, which is a more accurate
name. (It contains files that are not templates.) Use Jinaj's
PackageLoader and importlib.resources to access the data rather than
assuming the installed package is present in a file system, since this
appears to be the more Pythonic way of loading module resources.
