+++ src/mobu/data/tapqueryrunner/dp0.1/object-photometry.sql
