+++ src/mobu/data/tapqueryrunner/dp0.1/scan-magnitude.sql
