+++ src/mobu/data/tapqueryrunner/dp0.2-long-scans/time-series-several.sql
