+++ src/mobu/services/flock.py
@@ -12,15 +12,15 @@ from aiojobs import Scheduler
 from safir.datetime import current_datetime
 from structlog.stdlib import BoundLogger
 
+from ..exceptions import MonkeyNotFoundException
+from ..models.flock import FlockConfig, FlockData, FlockSummary
+from ..models.user import AuthenticatedUser, User, UserSpec
 from .business.base import Business
 from .business.jupyterjitterloginloop import JupyterJitterLoginLoop
 from .business.jupyterloginloop import JupyterLoginLoop
 from .business.jupyterpythonloop import JupyterPythonLoop
 from .business.notebookrunner import NotebookRunner
 from .business.tapqueryrunner import TAPQueryRunner
-from .exceptions import MonkeyNotFoundException
-from .models.flock import FlockConfig, FlockData, FlockSummary
-from .models.user import AuthenticatedUser, User, UserSpec
 from .monkey import Monkey
 
 __all__ = ["Flock"]
